issue #100 refactor: member id set recoil, localstorage
