+++ src/components/Landing/Sidebar/Profile/Profile.tsx
@@ -1,12 +1,21 @@
+import { useLayoutEffect } from "react";
+
+import { useRecoilState } from "recoil";
+
 import LogoutIcon from "@/assets/svg/ic-logout.svg?react";
 import ProfileImg from "@/assets/svg/profile-default.svg?react";
 
 import { Flex, Text } from "@/components/common";
 
+import { useGetMemberInfoMutation } from "@/hooks/api/useGetMemberInfoMutation";
 import { useLogoutMutation } from "@/hooks/api/useLogoutMutation";
+
+import { memberIdState } from "@/recoil/atoms/auth";
 
 import { getDefaultTextStyle } from "@/styles/getDefaultTextStyle";
 import { Theme } from "@/styles/Theme";
+
+import type { MemberInfoResponseType } from "@/types/auth";
 
 import {
 	layoutStyle,
@@ -17,6 +26,17 @@ import {
 
 const Profile = () => {
 	const { mutateLogOut } = useLogoutMutation();
+	const getMemberInfoMutation = useGetMemberInfoMutation();
+
+	const [memberId] = useRecoilState(memberIdState);
+
+	useLayoutEffect(() => {
+		getMemberInfoMutation.mutate(memberId, {
+			onSuccess: ({ result }: MemberInfoResponseType) => {
+				console.log(result);
+			},
+		});
+	}, [memberId]);
 
 	return (
 		<Flex css={layoutStyle}>
